"Application" class: JavaDoc updated.
Following the repurposing of "Person" class to "Application" class,
the JavaDoc of the class has become irrelevant.

JavaDoc inaccuracies need to be addressed so that developers reading
our code can have a better understanding.

Let's update the JavaDoc of "Application" class.

Going through and updating the JavaDoc in "Application" class will
suffice.

The JavaDoc for "tags" Data Field in this class is left as W.I.P;
Following the implementation of future tags this JavaDoc will be
updated.

The method isSamePerson() has also been refactored to
isSameApplication() however the implementation of this method will
need to be reviewed on a later date.