"Application" class: JavaDoc updated.
JavaDoc for isSameApplication() was missed in previous commit.

Let's update the JavaDoc for this method to reflect the change from
"Person" to "Application".

Implementation of the method has not been reviewed at this point.
isSameApplication() currently still checks the "name" data field to
determine if 2 applications are considered the same.