"NameContainsKeywordsPredicate" class: JavaDoc updated.
Following the repurposing of "Person" class to "Application" class,
the JavaDoc of the class has become irrelevant.

JavaDoc inaccuracies need to be addressed so that developers reading
our code can have a better understanding.

Let's update the JavaDoc of "NameContainsKeywordsPredicate" class.

Going through and updating the JavaDoc in
"NameContainsKeywordsPredicate" class will suffice.

Methods and purpose of this class have been left untouched. This
class is meant as part of the logic needed for the "find" command
in AB3.