"ReadOnlyAddressBook" class morphed to "ReadOnlyInternApplyMemory" class.
"ReadOnlyAddressBook" is a class from the AB3 code base. Currently,
it's named with the purpose of supporting AB3 development as an
address book.

Since we are morhping AB3 to InternApply to track Internship
applications, this class needs to be morphed as well.

Let's morph "ReadOnlyAddressBook" to "ReadOnlyInternApplyMemory"
class.

A simple refactoring will suffice.

InternApply is still able to succesfully run all previously
implemented tests (Test passed:248 of 248 tests)

"AddressBook" class that is found in the same folder as this class
has also been renamed to "InternApplyMemory" due to it's dependency
on "ReadOnlyAddressBook".