"InternApplyMemory" class: update methods.
"InterApplyMemory" is only a renaming of the old "AddressBook"
class.

The methods found in "InternApplyMemory" are badly named in the
context of InternApply.

Let's update the methods found in "InternApplyMemory".

A simple refactoring will suffice since we are still going to reuse
the method implementation from AB3 in InterApply.

InternApply is still able to succesfully run all previously
implemented tests (Test passed:248 of 248 tests).