AddressBookParser class: parseCommand() method now no longer recognizes the "edit" command
Currently, InternApply still accepts "edit" as a keyword.

For v1.2 of InternApply, this functionality is outside the scope of
the iteration.

Let's just disable the command by making AddressBookParser not
recognize "edit" as a keyword.

This will be done by simply commenting out the relevant switch-case
statement in the parseCommand() method in the AddressBookParser
class.

This will suffice since we do plan to implement this function in
future iterations of InternApply.

Non of the classes created to support the "edit" command has been
removed.

Instead, they will be updated in the future iterations of
InternApply.