"AddressBookParser" class morphed to "InternApplyParser" class.
"AddressBookParser" is a class from the AB3 code base. Currently,
it's named with the purpose of supporting AB3 development as an
address book.

Since we are morhping AB3 to InternApply to track Internship
applications, this class needs to be morphed as well.

Let's morph "AddressBookParser" to "InternApplyParser"
class.

A simple refactoring will suffice.

InternApply is still able to succesfully run all previously
implemented tests (Test passed:248 of 248 tests)

The implementation of methods in this class have not been changed.

As a by-product of this change, there has been a renaming of
"AddressBookParserTest" to "InternApplyParserTest" as well.