"JsonAdaptedPerson" class morphed to "JsonAdaptedApplication" class.
"JsonAdaptedPerson" is a class from the AB3 code base. Currently,
it's named with the purpose of supporting AB3 development as an
address book.

Since we are morphing AB3 to InternApply to track Internship
applications, this class needs to be morphed as well.

Let's morph "JsonAdaptedPerson" to "JsonAdaptedApplication" class.

A simple refactoring will suffice as well as the updating of
methods in "JsonAdaptedPerson" class.

InternApply is still able to succesfully run all previously
implemented tests (Test passed:248 of 248 tests)