"JsonInternApplyStorage" class: updated methods.
The class is still using naming conventions from AB3 codebase.

The method and variable names used in this class is irrelevant in
the context of InternApply.

Let's update the methods found in "JsonInternApplyStorage".

A simple refactoring will suffice since we are still going to reuse
the method implementation from AB3 in InternApply.

InternApply is still able to successfully run all previously
implemented tests (Test passed:248 of 248 tests).