"JsonSerializableAddressBook" class morphed to "JsonSerializableInternApply" class.
"JsonSerializableAddressBook" is a class from the AB3 code base.
Currently, it's named with the purpose of supporting AB3 development
as an address book.

Since we are morphing AB3 to InternApply to track Internship
applications, this class needs to be morphed as well.

Let's morph the class to "JsonSerializableInternApply" class.

A simple refactoring will suffice as well as the updating of
methods in "JsonAdaptedPerson" class.

InternApply is still able to succesfully run all previously
implemented tests (Test passed:248 of 248 tests).

Unable to rename field "persons" in line 22 of the class.
Renaming it to "applications" results in the application failing
the following testcases:
1. readAndSaveAddressBook_allInOrder_success()
2. addressBookReadSave()

Testcases can be found in:
src/test/java/seedu/address/storage/JsonInternApplyStorageTest.java
src/test/java/seedu/address/storage/StorageManagerTest.java
respectively.

Error could be related to how Json files are being handled in AB3
code base.