Add Job Title field to Application
Job Title field is required in job application to denote the job the
application is for.

Adding the Job Title field also requires the addition of the JobTitle
class, job title in the GUI, modifying the add command to take in
job title and modifications to the storage.

Let's also add 2 additional test cases for testing the add command with
the job title parameter as well as updating test data and test cases
to account for the Job Title field.

Note:
EditCommand is modified but does not allow editing job title.