Fix InterviewSlot comparison bug
Interview slot class equals() function was previously compared to an
instance of phone class resulting in errors in testcases.

Update to reflect correctly to compare value of interview slot.

A string comparison is more friendly due to parsing limitation
that truncated the default placeholder value - maximum value of
LocalDateTime class.