"HelpWindow" class: update USERGUIDE_URL.
Currently the USERGUIDE_URL still refers users to the AB3 UserGuide
website when users are looking for help.

Since we have already created out own website for InternApply's
UserGuide, we should be referring users to that link instead.

Let's update the USERGUIDE_URL so that it prints InternApply's
UserGuide website instead.

(Line 18) The attribute USERGUIDE_URL that contained the link to
AB3's UserGuide has been replaced with the link that takes users to
InternApply's UserGuide.

This does not affect any other part of the code or commands.