"PersonListPanel" class renamed to "ApplicationListPanel" class.
"PersonListPanel" is a class from the AB3 code base. Currently,
it's named with the purpose of supporting AB3 development as an
address book.

Since we are morphing AB3 to InternApply to track Internship
applications, this class needs to be morphed as well.

Let's start by renaming "PersonListPanel" to "ApplicationListPanel".

A simple refactoring will suffice for this.

InternApply is still able to succesfully run all previously
implemented tests (Test passed:251 of 251 tests)