PersonListPanel.fxml repurposed to ApplicationListPanel.fxml.
PersonListPanel.fxml is file from the AB3 code base. Currently,
it's named with the purpose of supporting AB3 development as an
address book, particularly the GUI component.

Since we are morphing AB3 to InternApply to track Internship
applications, this file is redundant as is.

Let's repurpose PersonListPanel.fxml to ApplicationListPanel.fxml.

A simple refactoring will suffice as well as the updating the
contents of the file.

Particularly, the "fx:id=" in Line 7 has been updated from
"personListView" to "applicationListView".

InternApply is still able to succesfully run all previously
implemented tests (Test passed:251 of 251 tests)

The class that makes reference to this fxml file,
ApplicationListPanel.java, has also been updated. (Line 17)