"ApplicationListPanel" class: updated @FXML attribute.
Currently, it still uses "personListView" as the @FXML attribute.

But this is no longer working resulting in an error when running
the application.

The source of the bug is determined to be the changing of the
"fx:id=" to "applicationListView" from "personListView" in the
ApplicationListPanel.fxml file.

Let's also rename the affected @FXML attribute in the
"ApplicationListPanel" class.

InternApply is still able to successfully run all previously
implemented tests (Test passed:251 of 251 tests)

This fixes the bug that prevents the application from opening.