PersonListCard.fxml repurposed to ApplicationListCard.fxml.
PersonListCard.fxml is file from the AB3 code base. Currently,
it's named with the purpose of supporting AB3 development as an
address book, particularly the GUI component.

Since we are morphing AB3 to InternApply to track Internship
applications, this file is redundant as is.

Let's repurpose PersonListCard.fxml to ApplicationListCard.fxml.

A simple refactoring will suffice as well as the updating the
contents of the file.

InternApply is still able to successfully run all previously
implemented tests (Test passed:251 of 251 tests)

The class that makes reference to this fxml file,
ApplicationCard.java, has also been updated. (Line 14)