"ApplicationCard" and "ApplicationListPanel" classes: updated JavaDoc and methods.
The class is still using naming conventions for methods from AB3
codebase and the JavaDocs reflects this.

The methods and JavaDocs are thus irrelevant in the context of
InternApply.

Let's update the methods and JavaDocs found in "ApplicationCard"
and "ApplicationListPanel".

InternApply is still able to successfully run all previously
implemented tests. (Test passed:251 of 251 tests)