"MainWindow" class: updated methods.
The class is still using naming conventions from AB3 codebase.

The method names used in this class is irrelevant in the context of
InternApply.

Let's update the methods found in "MainWindow".

A simple refactoring will suffice since we are still going to reuse
the method implementation from AB3 in InternApply.

InternApply is still able to successfully run all previously
implemented tests (Test passed:251 of 251 tests).