"UserPrefs" class: updated the field "addressBookFilePath".
Currently, "addressBookFilePath" specifies the name of the file the
app should search for on the user's local hard disk to load existing
data.

The file name as well as the name of the field is no longer relevant
in the context of InternApply.

Let's rename "addressBookFilePath" to "internApplyMemoryFilePath"
and also update the name of the json file from "addressbook.json" to
"internapplymemory.json".

These changes are reflected on Lines 17, 51, 56, 71 and 76.

InternApply now fails 2 of the 251 implemented tests
(Test passed:249 of 251 tests).

They are:
1. readUserPrefs_fileInOrder_successfullyRead()
2. readUserPrefs_extraValuesInFile_extraValuesIgnored()

This will need to be addressed in a later commit.