"JsonUserPrefsStorageTest" class: updated method. ExtraValuesUserPref.json & TypicalUserPref.json: changes made.
It's been identified that the two tests that are currently failing
have been implemented in the "JsonUserPrefsStorageTest" class.

However, these tests were created in the context of AB3 and as a
result, without modifications, they are failing.

Let's update these 2 specific tests to better reflect the changes
made in the previous commit.

In Line 76, we have changed the argument to
setInternApplyMemoryFilePath to the latest name
"internapplymemory.json".

Additionally, the 2 Json files that were created for the respective
tests have also been modified such that they now have the file path
"internApplyMemoryFilePath" that have the String value
"internapplymemory.json".

This modification better reflects the behavior of the app in the
current iteration.

InternApply is now able to successfully run all implemented tests
(Test passed:251 of 251 tests).

This commit directly addresses the issues #69 and #70 in the tP
repo.