"ClearCommand" class: updated class.
Currently, the message that is being printed when using the "clear"
command is still in the context of an address book being cleared.

Since our application now deals with internship applications, the
message is no longer relevant.

Let's change the messages for the "ClearCommand" class.

All the relevant attributes in the class will be changed to reflect
the action of "clearing InternApply".

InternApply is still able to successfully run all previously
implemented tests (Test passed:251 of 251 tests)

Minor changes have been made to the JavaDoc of this class as well.