"ExitCommand" class: updated class.
Currently, the acknowledgement message of the "exit" command is
still written in the context of an address book being exited.

Since our application now deals with internship applications, the
message is no longer relevant.

Let's change the message for the "ExitCommand" class.

All the relevant attribute in the class will be changed to reflect
the action of "exiting from InternApply".

InternApply is still able to successfully run all previously
implemented tests (Test passed:251 of 251 tests)