"FindCommand" class: updated class.
Currently, the messages that are being printed when using the
"find" command are still in the context of an address book being
used.

Since our application now deals with internship applications, these
messages are no longer relevant.

Let's change the messages for the "FindCommand" class.

All the relevant attributes in the class will be changed to reflect
the action of "finding applications that match the given keyword".

InternApply is still able to successfully run all previously
implemented tests (Test passed:251 of 251 tests)

Minor changes have been made to the JavaDoc of this class as well.