"PersonUtil" class morphed to "ApplicationUtil" class.
"PersonUtil" is a class from the AB3 code base that is being used
for testing purposes.

Since we are morphing AB3 to InternApply to track Internship
applications, we should morph this class under the context of an
application.

Let's morph "PersonUtil" to "ApplicationUtil" class.

A simple refactoring will suffice as well as the updating of
methods and JavaDoc in the "ApplicationUtil" class.

InternApply is still able to successfully run all previously
implemented tests (Test passed:251 of 251 tests)