"EditApplicationDescriptorTest" JUnit class: update naming convention.
"EditApplicationDescriptorTest" is a JUnit class from the AB3 code
base.

Since we are morphing AB3 to InternApply to track Internship
applications, this test should reflect the change from person to
application.

Let's update the naming conventions used in this JUnit test.

InternApply is able to successfully run all tests in this JUnit
class.