"EditCommand" class: modify implementation.
The way the edit command works currently is for a new application
to be be instantiated within the "EditCommand" that will the replace
the existing application at the given index.

This leads to a behavior in InternApply that causes it to wipe all
existing tags of an application when editing the tags of an
application.

Since we have expanded on the functionality of edit to allow users
to also edit application status and priority tag, this previously
acceptable behavior has become a problem akin to a bug.

We want users to be able to use the edit command to edit the
generic tags, Application Status Tag or Priority Tag of existing
applications without having to worry about reentering existing tags
that aren't being edited.

Let's modify the behavior of the edit command such that it knows
what tags are being modified and only overrides those tags in the
existing application while remembering to bring over the other tags
to the newly created application.

We will create a new method getTags() that now takes in the entire
Set<Tag> applicationToEditTags. If all three tags were left
unedited, we will simply copy the existing tags over to the new
application.

Otherwise, we want to reconstruct the Set<Tag> object such that it
contains both new tags and also remembering old tags that were not
touched.

For every tag (divided by prefix) we will first check if edits were
made by the user. If so, we will copy the new edits and override the
old ones. This behavior follows the current behavior of InternApply.

If InternApply cannot detect any edits made by the user, InternApply
will simply copy over the tags from the existing application.

In doing so, we ensure that InternApply only overrides relevant tags
that have been selected for edits by the user.

Due to the changes made, 1 of the test cases in the
"EditCommandTest" class is failing. A commit in the future will be
made to address this.