Create "ReminderCommand" class. Update InternApplyParser#parseCommand().
Currently there is no way for a user to easily check what are the
upcoming interviews that they have.

We want to provide this as a feature by implementing a new
"reminder" command.

Let's start by creating a "ReminderCommand" class that extends the
"Command" class. Then update InternApplyParser#parseCommand() method
to recognize "reminder" as the keyword to call the command.

We will create a "ReminderCommand" class that extends "Command"
class.

Next, we include a case where "reminder" is detected as an input in
InternApplyParser#parseCommand().

InternApply is still able to successfully run all previously
implemented tests (Test passed:264 of 264 tests).