Modify CommandResult constructor.
Currently, CommandResult constructor takes in two Boolean values
that represents if the CommandResult should show help or exit the
program.

"help" command would create a CommandResult that should show help
but not exit.

"exit" command would create a CommandResult that should exit but not
show help.

All other commands would create a CommandResult that should not show
help or exit.

"show help" here means to create the help window.

Since we are creating "reminder" command in a similar vein as "help"
command where we would like to create a reminder window we need to
recreate this behavior of "show reminder".

Let's modify the constructor for "CommandResult" class.

We will modify it such that it takes in a third Boolean named
showReminder that is only true when ReminderCommand#execute() is
called that creates a CommandResult.

InternApply is still able to successfully run all previously
implemented tests (Test passed:264 of 264 tests).

All instances of CommandResult being created have been updated as
such:

"help" command would create a CommandResult that should show help
but not show reminder or exit.

"exit" command would create a CommandResult that should exit and not
show help or reminder.

All other commands would create a CommandResult that should not show
help, show reminder or exit.

"reminder" command would create a CommandResult that should show
reminder but not show help or exit.

Additionally, all test cases that instantiates a CommandResult
object have been updated accordingly to the above mentioned
behavior.