Implement Ui for "reminder" command.
Currently, calling the "reminder" command only returns a response
with text.

We want it such that a window with a list of upcoming interviews
show up in a pop-up window.

Let's implement the Ui support for the "reminder" command.

As a start, we will first be making a copy of the Ui support for
"help" command.

We will be building on this in future commits.

InternApply now runs successfully run all 265 tests
(Test passed:265 of 265 tests).