Modified the contents of the reminder window for the "reminder" command.
Currently, the reminder window only contains a short placeholder
message.

We want it such that a window with a list of applications with
upcoming interviews show up in the pop-up window.

Let's next make it such that the reminder window is populated with
a list of applications.

InternApply now runs successfully run all 265 tests
(Test passed:265 of 265 tests).