Add a way to customize startup behavior of InternApply.
Currently, on startup only the Main Window is created and open.

We want to be able to customize additional behaviors of InternApply
on startup.

Let's implement a way for us to run a code on startup before the
control is handed to the user and the application starts accepting
user input.

We can implement a MainWindow#init() method that is a void method
that just runs the specificed code.

We modify UiManager#start() method such that it now calls
MainWindow#init() before terminating and handing the control back
to MainApp#start().

InternApply now runs successfully run all 265 tests
(Test passed:265 of 265 tests).

For now, MainWindow#init() has been implemented such that it calls
"reminder" command.

Thus, on startup, the both the Main Window and Reminder Window will
be launched before control is given to the user.