Update SampleDataUtil#getSampleApplication() method.
Currently, the method is being used to create Sample Data that will
be used to populate InternApply's application list.

This method is only called if an existing internapplymemory.json
file is not detected in the data folder.

The Sample Data is a hard coded set of applications. This means that
the Interview Slots created are generally going to be irrelevant at
some point.

Let's implement a way to generate Interview Slots that are more
useful.

To do this, we want to check for the LocalDateTime.now() of the
local machine and use that to generate our Interview Slots.

This ensures that no matter which user is using InternApply or when
they are using InternApply, the Sample Data generated (if needed)
will be relevant relative to the actual current local date time of
his/her machine.