Update ParserUtil#parseTags() method.
Currently, the method simply checks if the Collection of Strings
passed to it is null or not.

This means that even a Collection of empty Set<String> would pass
this check as a valid input.

Let's change this since an empty Set<String> not be meaningful and
should be rejected.

We simply include an if-else statement that determines if the
Collection passed in is an empty Set<String>.

EditCommandParserTest#parse_resetTags_success() has been changed to
parse_resetTags_failure() since the behavior that it is testing for
no longer exists and "t/" as an empty input is now invalid.